requirements.txt
# ============================================================
# FILE: requirements.txt
# ============================================================

opencv-python==4.8.1.78
numpy==1.24.3
PyYAML==6.0.1
flask==3.0.0
paho-mqtt==1.6.1
ultralytics==8.1.0
tensorflow==2.15.0
tflite-runtime==2.14.0
Pillow==10.1.0
python-dotenv==1.0.0
twilio==8.10.0
